+++ src/main/scala-2.11/markets/auctions/continuous/ContinuousDoubleAuction.scala
@@ -13,14 +13,17 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package markets.auctions
+package markets.auctions.continuous
 
+import markets.auctions.{DoubleAuction, Fill}
 import markets.tradables.orders.ask.AskOrder
 import markets.tradables.orders.bid.BidOrder
 
 
-trait PeriodicDoubleAuction[A <: AskOrder, B <: BidOrder] extends DoubleAuction[A, B] {
+trait ContinuousDoubleAuction[A <: AskOrder, B <: BidOrder] extends DoubleAuction[A, B] {
 
-  def fill(): Iterable[Fill]
+  def fill(order: A): Option[Fill]
+
+  def fill(order: B): Option[Fill]
 
 }
