+++ src/main/scala-2.11/markets/auctions/periodic/PeriodicDoubleAuction.scala
@@ -13,16 +13,15 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package markets.auctions
+package markets.auctions.periodic
 
+import markets.auctions.{DoubleAuction, Fill}
 import markets.tradables.orders.ask.AskOrder
 import markets.tradables.orders.bid.BidOrder
 
 
-trait ContinuousDoubleAuction[A <: AskOrder, B <: BidOrder] extends DoubleAuction[A, B] {
+trait PeriodicDoubleAuction[A <: AskOrder, B <: BidOrder] extends DoubleAuction[A, B] {
 
-  def fill(order: A): Option[Fill]
-
-  def fill(order: B): Option[Fill]
+  def fill(): Iterable[Fill]
 
 }
